backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
opencv-python==4.8.1.78
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
marshmallow==3.20.1

Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0